Clean up API response
