+++ db/project.ts
@@ -309,45 +309,3 @@ export async function updateProject(
     console.error(error)
   }
 }
-
-export async function listProjectsPaginated(
-  supabase: SupabaseClient,
-  before?: string | null,
-  causes?: string[]
-) {
-  let query = supabase
-    .from('projects')
-    .select(
-      `
-      title, 
-      id, 
-      created_at, 
-      creator,
-      slug,
-      blurb,
-      description,
-      stage,
-      funding_goal,
-      min_funding,
-      type,
-      profiles!projects_creator_fkey(username),
-      txns(amount),
-      causes(title, slug)
-    `
-    )
-    .neq('stage', 'hidden')
-    .neq('stage', 'draft')
-    .order('created_at', { ascending: false })
-    .limit(100)
-
-  if (before) {
-    query = query.filter('id', 'lt', before)
-  }
-
-  if (causes && causes.length > 0) {
-    query = query.filter('causes.slug', 'in', `(${causes.join(',')})`)
-  }
-
-  const { data } = await query.throwOnError()
-  return data
-}
