+++ package.json
@@ -58,6 +58,7 @@
     "eslint-config-next": "^14.0.4",
     "lodash": "^4.17.21",
     "next": "14.2.5",
+    "node-html-markdown": "^1.3.0",
     "rc-slider": "^10.1.1",
     "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
